Miscellaneous fixes.

+++ libs/execution_context/ebpf_maps.c
@@ -2996,8 +2996,8 @@ ebpf_map_get_next_key_and_value_batch(
 
         memcpy(key_and_value + output_length + key_size, next_value, value_size);
 
-        if (flags & EBPF_MAP_FIND_FLAG_DELETE) {
-            // If the caller requested deletion, delete the entry.
+        if ((flags & EBPF_MAP_FIND_FLAG_DELETE) && (previous_key != NULL)) {
+            // If the caller requested deletion, delete the previous entry.
             result = table->delete_entry(map, previous_key);
             if (result != EBPF_SUCCESS) {
                 EBPF_LOG_MESSAGE_UINT64(
@@ -3021,8 +3021,7 @@ ebpf_map_get_next_key_and_value_batch(
 
     if ((flags & EBPF_MAP_FIND_FLAG_DELETE) && (previous_key != NULL) && (output_length != 0)) {
         // If the caller requested deletion, delete the last entry.
-        ebpf_result_t delete_result =
-            table->delete_entry(map, previous_key);
+        ebpf_result_t delete_result = table->delete_entry(map, previous_key);
         if (delete_result != EBPF_SUCCESS) {
             EBPF_LOG_MESSAGE_UINT64(
                 EBPF_TRACELOG_LEVEL_ERROR, EBPF_TRACELOG_KEYWORD_MAP, "Failed to delete last entry", delete_result);
